'Improve' logout behavior
After logout, subsequent visits to /secure would initiate Google
OAuth flow as expected, but immediately succeed and yield a new
session cookie token. This is actually how many apps choose to
work, since if the user is still logged into Google this makes
for a seamless experience. In this case, I have chosen to add
a bit of friction via the select_account oauth prompt parameter
so it is visually obvious what is happening to the user after
logout and subsequent revisit to /secure.