Implements requirement #6
Ensures that JWTs POSTed to either /insecure or /secure forms that
validate against a known public EC key will render with a red table
row background in the /secure view. This is how I chose to interpret
the instructions, since posting a signed JWT to /insecure and then
browsing /secure only to find the entry is not highlighted might be
suprising. (principle of least surprise)

adds helpers to generate signed JWTs for testing.