Deployment to GCP Compute Engine with COS via Terraform
To keep things simple, this is a straightforard 'lift and shift' of
the docker-compose setup that runs locally. We'll use the e2-micro
instance type to keep it in the free tier, and also an ephemeral ip
address, since static ips have a monthly fee. I selected the COS
(container optimized OS) which has docker pre-installed.

I've chosen to re-use the top-level .env file, and pass these along
through terraform invocation, and into the startup script, to avoid
duplication of secrets in tfvars files.